requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pyee==7.0.4
pymongo==3.11.3
python-dotenv==0.15.0
slack-sdk==3.4.0
slackeventsapi==2.2.1
Werkzeug==1.0.1
